Refactor header menu to remove redundant information
